requirements.txt
# ──────────────────────────────────────────────
# Core Vanna 2.0
# ──────────────────────────────────────────────
vanna>=2.0.0
openai
# ──────────────────────────────────────────────
# LLM Provider — Groq (Option B)
# OpenAI-compatible; set GROQ_API_KEY in .env
# ──────────────────────────────────────────────
groq

# ──────────────────────────────────────────────
# API Framework
# ──────────────────────────────────────────────
fastapi
uvicorn[standard]

# ──────────────────────────────────────────────
# Data & Visualisation
# ──────────────────────────────────────────────
plotly
pandas

# ──────────────────────────────────────────────
# Utilities
# ──────────────────────────────────────────────
python-dotenv
